feat: 이미지 캡처 오류 수정
